Add rack middleware
This will look at connection_info in each ParticipantService to check for a "domain" key
that matches the incoming response from a request to a third-party.

Additionally, the middleware may be setup manually or you can use CloseEncounters.auto_contact!
to turn on request tracking automatically in the Rails middleware stack.